DM-153 Refactored Navbar feature
